refactor: :hammer: 공통으로 사용되는 state와 함수 write 페이지로 이동
